feat(stage-wizard): sortable wizard card
