--- packages/compass-aggregations/src/components/pipeline-builder-workspace/index.js
@@ -1,2 +1,0 @@
-import PipelineBuilderWorkspace from './pipeline-builder-workspace';
-export default PipelineBuilderWorkspace;
