+++ packages/compass-aggregations/src/components/pipeline-builder-workspace/index.tsx
